Move pandas/dask/xarray pipeline logic to data_libraries package
